requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.28.2
python-dotenv==1.0.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
redis==4.5.5
pandas==2.0.1
scikit-learn==1.2.2